[Setting] 문제풀이 세팅
